# api/apps/api/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MarxanCloud API",
    "description": "MarxanCloud is a conservation planning platform.",
    "version": "0.4.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
        }
    },
    "schemas": {
...
      "ScenarioType": {
        "type": "string",
        "enum": ["marxan", "marxan-with-zones"]
      },
      "JobStatus": {
        "type": "string",
        "enum": ["created", "running", "done", "failure"]
      },
      "Scenario": {
        "type": "object",
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdByUser": {
            "$ref": "#/components/schemas/User"
          },
          "lastModifiedAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/ScenarioType"
          },
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "wdpaIucnCategories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "protectedAreaFilterByIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "wdpaThreshold": {
            "type": "object"
          },
          "numberOfRuns": {
            "type": "number"
          },
          "boundaryLengthModifier": {
            "type": "number"
          },
          "metadata": {
            "type": "object"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "parentScenarioId": {
            "type": "string"
          },
          "parentScenario": {
            "$ref": "#/components/schemas/Scenario"
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          }
        },
        "required": [
          "createdAt",
          "createdByUser",
          "lastModifiedAt",
          "id",
          "name",
          "type",
          "project",
          "status",
          "users"
        ]
      },
      "Project": {
        "type": "object",
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdByUser": {
            "$ref": "#/components/schemas/User"
          },
          "lastModifiedAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "organization": {
            "$ref": "#/components/schemas/Organization"
          },
          "adminAreaLevel1Id": {
            "type": "string"
          },
          "adminAreaLevel2Id": {
            "type": "string"
          },
          "planningUnitGridShape": {
            "type": "string"
          },
          "planningUnitAreakm2": {
            "type": "number"
          },
          "extent": {
            "type": "object"
          },
          "bbox": {
            "type": "object"
          },
          "metadata": {
            "type": "object"
          },
          "scenarios": {
            "$ref": "#/components/schemas/Scenario"
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          },
          "planningAreaId": {
            "type": "string",
            "description": "ID of Country / Gid1 / Gid2 of project's area"
          },
          "planningAreaName": {
            "type": "string",
            "description": "Display name of Country / Gid1 / Gid2 of project's area"
          }
        },
        "required": [
          "createdAt",
          "createdByUser",
          "lastModifiedAt",
          "id",
          "name",
          "organization",
          "adminAreaLevel1Id",
          "adminAreaLevel2Id",
          "planningUnitGridShape",
          "planningUnitAreakm2",
          "bbox",
          "users"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "object"
          },
          "fname": {
            "type": "object"
          },
          "lname": {
            "type": "object"
          },
          "avatarDataUrl": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "isActive": {
            "type": "boolean"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "projects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Project"
            }
          },
          "scenarios": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Scenario"
            }
          }
        },
        "required": ["email", "isActive", "isDeleted", "projects", "scenarios"]
      },
      "Organization": {
        "type": "object",
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdByUser": {
            "$ref": "#/components/schemas/User"
          },
          "lastModifiedAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "projects": {
            "$ref": "#/components/schemas/Project"
          }
        },
        "required": [
          "createdAt",
          "createdByUser",
          "lastModifiedAt",
          "id",
          "name"
        ]
      },
...
      "UpdateUserPasswordDTO": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "required": ["currentPassword", "newPassword"]
      },
      "JSONAPIUserData": {
        "type": "object",
        "properties": {
          "type": {
            "type": "object"
          },
          "id": {
            "type": "string"
          },
          "attributes": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": ["type", "id", "attributes"]
      },
      "UserResult": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/JSONAPIUserData"
          }
        },
        "required": ["data"]
      },
      "UpdateUserDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "object"
          },
          "fname": {
            "type": "object"
          },
          "lname": {
            "type": "object"
          },
          "password": {
            "type": "string"
          },
          "avatarDataUrl": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "summary": "",
        "description": "Find all users",
        "parameters": [
          {
            "name": "include",
            "required": false,
            "in": "query",
            "description": "A comma-separated list of relationship paths. Allows the client to customize which related resources should be returned. Allowed values are: `projects`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "An array of filters (e.g. `filter[keyA]=<value>&filter[keyB]=<value1>,<value2>...`). Allows the client to request for specific filtering criteria to be applied to the request. Semantics of each set of filter key/values and of the set of filters as a whole depend on the specific request.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "A comma-separated list that refers to the name(s) of the fields to be returned. An empty value indicates that all fields will be returned (less any fields specified as `omitFields`).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "omitFields",
            "required": false,
            "in": "query",
            "description": "A comma-separated list that refers to the name(s) of fields to be omitted from the results. This could be useful as a shortcut when a specific field such as large geometry fields should be omitted, but it is not practical or not desirable to explicitly whitelist fields individually. An empty value indicates that no fields will be omitted (although they may still not be present in the result if an explicit choice of fields was provided via `fields`).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "A comma-separated list of fields of the primary data according to which the results should be sorted. Sort order is ascending unless the field name is prefixed with a minus (for descending order).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page[size]",
            "required": false,
            "in": "query",
            "description": "Page size for pagination. If not supplied, pagination with default page size of 25 elements will be applied.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page[number]",
            "required": false,
            "in": "query",
            "description": "Page number for pagination. If not supplied, the first page of results will be returned.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "disablePagination",
            "required": false,
            "in": "query",
            "description": "If set to `true`, pagination will be disabled. This overrides any other pagination query parameters, if supplied.",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "The current user does not have suitable permissions for this request."
          }
        },
        "tags": ["User"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/users/me/password": {
      "patch": {
        "operationId": "UsersController_updateOwnPassword",
        "summary": "",
        "description": "Update the password of a user, if they can present the current one.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserPasswordDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResult"
                }
              }
            }
          }
        },
        "tags": ["User"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/users/me": {
      "patch": {
        "operationId": "UsersController_update",
        "summary": "",
        "description": "Update a user.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResult"
                }
              }
            }
          }
        },
        "tags": ["User"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "UsersController_userMetadata",
        "summary": "",
        "description": "Retrieve attributes of the current user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "The current user does not have suitable permissions for this request."
          }
        },
        "tags": ["User"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "UsersController_deleteOwnUser",
        "summary": "",
        "description": "Mark user as deleted.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "403": {
            "description": ""
          }
        },
        "tags": ["User"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  }
}